gatherers/tiktok/requirements.txt
kafka-python
python-dotenv
apify-client
requests
